// app/lib/services/lyrics/types/genius.types.ts
export interface Meta {
  status: number
}

export interface ResponseReferentsRange {
  content: string
}

export interface ResponseReferentsAnnotatableClientTimestamps {
  updated_by_human_at: number
  lyrics_updated_at: number
}

export interface ResponseReferentsAnnotatable {
  api_path: string
  client_timestamps: ResponseReferentsAnnotatableClientTimestamps
  context: string
  id: number
  image_url: string
  link_title: string
  title: string
  type: string
  url: string
}

export interface ResponseReferentsAnnotationsBodyDomChildren {
  tag: string
  children: string[]
}

export interface ResponseReferentsAnnotationsBodyDom {
  tag: string
  children: ResponseReferentsAnnotationsBodyDomChildren[]
}

export interface ResponseReferentsAnnotationsBody {
  plain: string
}

export interface ResponseReferentsAnnotationsCurrentUserMetadataInteractions {
  cosign: boolean
  pyong: boolean
  vote?: unknown
}

export interface ResponseReferentsAnnotationsCurrentUserMetadata {
  permissions?: unknown[]
  excluded_permissions: string[]
  interactions: ResponseReferentsAnnotationsCurrentUserMetadataInteractions
  iq_by_action: object
}
// ...
export interface ResponseReferentsAnnotationsAuthorsUser {
  api_path: string
  avatar: ResponseReferentsAnnotationsAuthorsUserAvatar
  header_image_url: string
  human_readable_role_for_display: string
  id: number
  iq: number
  login: string
  name: string
  role_for_display: string
  url: string
  current_user_metadata: ResponseReferentsAnnotationsAuthorsUserCurrentUserMetadata
}

export interface ResponseReferentsAnnotationsAuthors {
  attribution: number
  pinned_role: string
  user: ResponseReferentsAnnotationsAuthorsUser
}

export interface ResponseReferentsAnnotations {
  api_path: string
  body: ResponseReferentsAnnotationsBody
  comment_count: number
  community: boolean
  custom_preview?: unknown
  has_voters: boolean
  id: number
  pinned: boolean
  share_url: string
  source?: unknown
  state: string
  url: string
  verified: boolean
  votes_total: number
  current_user_metadata: ResponseReferentsAnnotationsCurrentUserMetadata
  authors: ResponseReferentsAnnotationsAuthors[]
  cosigned_by?: unknown[]
  rejection_comment?: unknown
  verified_by: ResponseReferentsAnnotationsAuthorsUser
}

export interface ResponseReferents {
  _type: string
  annotator_id: number
  annotator_login: string
  api_path: string
  classification: string
  fragment: string
  id: number
  is_description: boolean
  path: string
  range: ResponseReferentsRange
  song_id: number
  url: string
  verified_annotator_ids?: unknown[]
  annotatable: ResponseReferentsAnnotatable
  annotations: ResponseReferentsAnnotations[]
}

export interface Response {
  referents: ResponseReferents[]
}
// ...
export interface SearchResponse {
  meta: Meta
  response: Response
}